+++ source/sound.cpp
@@ -1810,7 +1810,7 @@ static void soundEvent_u16_parallel(uint32_t address[])
 
 			default:
 				{
-					int gb_addr[2]	= {address[i] & ~1, address[i] | 1};
+					int gb_addr[2]	= {(int) (address[i] & ~1), (int) (address[i] | 1)};
 					uint32_t address_array[2] = {address[i] & ~ 1, address[i] | 1};
 					uint8_t data_array[2] = {0};
 					gba_to_gb_sound_parallel(&gb_addr[0], &gb_addr[1]);
@@ -1831,7 +1831,7 @@ static void gba_pcm_fifo_timer_overflowed( unsigned pcm_idx )
 		if ( pcm[pcm_idx].count <= 16 )
 		{
 			// Not filled by DMA, so fill with 16 bytes of silence
-			int reg = pcm[pcm_idx].which ? FIFOB_L : FIFOA_L;
+			uint32_t reg = (uint32_t) (pcm[pcm_idx].which ? FIFOB_L : FIFOA_L);
 
 			uint32_t address_array[8] = {reg, reg+2, reg, reg+2, reg, reg+2, reg, reg+2};
 			soundEvent_u16_parallel(address_array);
@@ -1926,7 +1926,7 @@ void soundEvent_u16(uint32_t address, uint16_t data)
 
 		default:
 			{
-				int gb_addr[2]	= {address & ~1, address | 1};
+				int gb_addr[2]	= {(int) (address & ~1), (int) (address | 1)};
 				uint32_t address_array[2] = {address & ~ 1, address | 1};
 				uint8_t data_array[2] = {(uint8_t)data, (uint8_t)(data >> 8)};
 				gba_to_gb_sound_parallel(&gb_addr[0], &gb_addr[1]);
